feature-StockBot/requirements.txt
flask==3.1.0
black==25.1.0
flake8==7.1.2
flask_sqlalchemy==3.1.1
python-dotenv==1.0.0
alpaca-trade-api==3.0.0
mysql-connector-python==8.0.33
